Fixed CUDA Coda and adjusted tcPDA.m for GPU and CPU calculation.
git-svn-id: svn://10.153.116.2/PAM@314 cb18dcf6-275f-e045-ae08-2e8c34a2df07

+++ tcPDA C-Code/C Files/src/sys/tcpda_cuda.h
@@ -16,11 +16,11 @@ typedef struct tcpda_data_struct {
 	float *fgg;
 	float *fgr;
 	
-	float *NBGbb;       /* NBGbb = numel(BGbb)-1 */
-	float *NBGbg;
-	float *NBGbr;
-	float *NBGgg;
-	float *NBGgr;
+	int *NBGbb;       /* NBGbb = numel(BGbb)-1 */
+	int *NBGbg;
+	int *NBGbr;
+	int *NBGgg;
+	int *NBGgr;
 	
 	float *BGbb;       /* Background corrections */
 	float *BGbg;
